Fix: 요식업자 프로필 삭제, account 앱에 몰아넣기
