+++ account/admin.py
@@ -4,3 +4,4 @@ from .models import *
 # Register your models here.
 admin.site.register(User)
 admin.site.register(Tag)
+admin.site.register(ConsultantProfile)
