+++ myProfile/admin.py
@@ -2,5 +2,3 @@ from django.contrib import admin
 from .models import *
 
 # Register your models here.
-admin.site.register(RestaurantProfile)
-admin.site.register(ConsultantProfile)
